requirements.txt
flask
flask-cors
python-docx
youtube-transcript-api
gunicorn
